feat:Roles and Permission enhanced.

+++ app/Http/Controllers/ClusterController.php
@@ -17,6 +17,9 @@ class ClusterController extends Controller
      */
     public function index()
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin|Staff|Student')) {
+            return redirect('/')->with('error', 'Unauthorised to access this page.');
+        }
         $clusters = Cluster::paginate(6);
         return view('clusters.index', compact(['clusters', ]));
     }
@@ -26,6 +29,9 @@ class ClusterController extends Controller
      */
     public function create()
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin')) {
+            return redirect('/')->with('error', 'Unauthorised to create clusters');
+        }
         $units = Unit::all();
         return view('clusters.create', compact(['units']));
     }
@@ -64,6 +70,9 @@ class ClusterController extends Controller
 
     public function show(string $id)
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin|Staff|Student')) {
+            return redirect('/')->with('error', 'Unauthorised to view clusters details.');
+        }
         $cluster = Cluster::find($id);
 
         if ($cluster) {
@@ -94,6 +103,10 @@ class ClusterController extends Controller
      */
     public function edit(string $id)
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin')) {
+            return redirect('/')->with('error', 'Unauthorised to edit clusters.');
+        }
+
         $cluster = Cluster::findOrFail($id);
         $units = Unit::all();
 
@@ -131,6 +144,9 @@ class ClusterController extends Controller
      */
     public function destroy(Cluster $cluster)
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin')) {
+            return redirect('/')->with('error', 'Unauthorised to delete clusters.');
+        }
         $cluster->delete();
         return redirect()->route('clusters.index')
             ->with('success', 'Cluster deleted successfully');
