+++ app/Http/Controllers/PackageController.php
@@ -15,6 +15,10 @@ class PackageController extends Controller
      */
     public function index()
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin|Staff')) {
+            return redirect('/')->with('error', 'Unauthorised to access this page.');
+        }
+
         $packages = Package::paginate(10);
         return view('packages.index', compact(['packages']));
     }
@@ -24,6 +28,10 @@ class PackageController extends Controller
      */
     public function create()
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin')) {
+            return redirect('/')->with('error', 'Unauthorised to create package!');
+        }
+
         $courses = Course::all();
         $package = new Package();
         return view('packages.create',compact(['courses', 'package' ]));
@@ -33,8 +41,7 @@ class PackageController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-
-            {
+    {
         $validated = $request->validate([
             'national_code' => ['required', 'string', 'size:3',  'regex:/^[A-Z]/'],
             'title' => ['required', 'min:5', 'max:255', 'string',],
@@ -58,9 +65,11 @@ class PackageController extends Controller
      */
     public function show(Package $package)
     {
-        {
+        if (!auth()->user()->hasRole('Super Admin|Admin|Staff')) {
+            return redirect('/')->with('error', 'Unauthorised to view package!');
+        }
 
-            $courses = $package->courses;
+        $courses = $package->courses;
             if ($package) {
                 return view('packages.show', compact(['package', 'courses']))
                     ->with('success', 'Package found')
@@ -69,13 +78,17 @@ class PackageController extends Controller
 
             return redirect(route('packages.index'))
                 ->with('warning', 'Package not found');
-        }    }
+    }
 
     /**
      * Show the form for editing the specified resource.
      */
     public function edit(Package $package)
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin')) {
+            return redirect('/')->with('error', 'Unauthorised to edit package!');
+        }
+
         $courses = Course::all();
         return view('packages.edit', compact('package', 'courses'));
     }
@@ -113,6 +126,10 @@ class PackageController extends Controller
      */
     public function destroy(Package $package)
     {
+        if (!auth()->user()->hasRole('Super Admin|Admin')) {
+            return redirect('/')->with('error', 'Unauthorised to delete package!');
+        }
+
         $package->delete();
         return redirect()->route('packages.index')
             ->with('success', 'Package deleted successfully');
